添加新指令，修复部分错误

+++ src/plugins/group/recall.py
@@ -9,7 +9,6 @@ from litebot_utils.rule import is_group_admin
 from src.plugins.menu.manager import MatcherData
 
 recall = on_message(
-    rule=is_group_admin,
     state=MatcherData(
         rm_name="撤回消息",
         rm_desc="用机器人撤回一条消息",
@@ -23,6 +22,7 @@ async def _(event: GroupMessageEvent, bot: Bot, matcher: Matcher):
         return
     if not event.reply:
         await matcher.finish("请回复消息选择撤回")
-
+    if not await is_group_admin(event, bot):
+        return
     await bot.delete_msg(message_id=event.reply.message_id)
     await matcher.finish("已尝试撤回消息！")
